Revert "some size_t code tidyups to remove windows warnings etc"
This reverts commit d7ac9b1af5b606235865b72779935d8974da1bb5.

+++ include/ngl/ShaderLib.h
@@ -165,7 +165,7 @@ public :
   /// @brief get the number of shaders stored
   /// @returns the number of shaders
   //----------------------------------------------------------------------------------------------------------------------
-  size_t getNumShaders() const
+  unsigned int getNumShaders() const
   {
     return m_shaders.size();
   }
